update install default deduction
in newer versions of cmake, there is a special variable we can use to
deduce if the user defined a path or not. This allows the project
default to actually be a default and the user can change the install
path to happen to correspond to the cmake default if they so wish.