applied updates from PF's branch of GBL on GitHub
I generated a patch file comparing PF's branch to the release it was
based on.

  git checkout pf/master
  git diff --patch -- cpp/include cpp/src > pbutti.diff

Then I switch to this branch and applied this patch with a three-way
merge so I could manually resolve the conflicts.

  git checkout java-integration
  git apply -3 pbutti.diff

Now I have committed this WITHOUT EVEN CHECKING COMPILATION. I will need
more expert help from PF to deduce which changes are important and which
ones are not.